+++ packages/moonstone/internal/Picker/Picker.js
@@ -55,7 +55,9 @@ const now = function () {
 const emulateMouseEventsTimeout = 175;
 
 // Set-up event forwarding
-const forwardClick = forward('onClick'),
+const forwardBlur = forward('onBlur'),
+	forwardClick = forward('onClick'),
+	forwardFocus = forward('onFocus'),
 	forwardKeyDown = forward('onKeyDown'),
 	forwardMouseDown = forward('onMouseDown'),
 	forwardMouseUp = forward('onMouseUp'),
@@ -359,13 +361,17 @@ const Picker = class extends React.Component {
 		}
 	}
 
-	handleBlur = () => {
+	handleBlur = (ev) => {
+		forwardBlur(ev, this.props);
+
 		this.setState({
 			active: false
 		});
 	}
 
-	handleFocus = () => {
+	handleFocus = (ev) => {
+		forwardFocus(ev, this.props);
+
 		this.setState({
 			active: true
 		});