+++ packages/moonstone/CHANGELOG.md
@@ -8,6 +8,7 @@ The following is a curated list of changes in the Enact moonstone module, newest
 
 - `moonstone/Panels.Panel` behavior for setting focus after render
 - `moonstone/VirtualList.VirtualGridList` showing empty items when keep adding items dynamically
+- `moonstone/Picker` to marquee on focus once again
 
 ## [1.0.0-beta.4] - 2017-03-10